If video offset negative, video starts at end
The last commit made it so a negative
time runoff does not occur when the
data is ahead of the video.  This
commit fixes the case when the video offset
is negative.  When the offset was negative,
the video was jumping back which actually
put it at the end of the video minus the
offset.  I thought I had tested negative
offsets, but my test wasn't what I thought
it was!