Add toggle between single and multiview
This adds the little clickable white box
to each video in the grid of videos.  When
one clicks the white box, it toggles
between single and multi video view.

I could not simply have the user click
on a video to expand it since I couldn't
get the mouse click event from the mpv
container widget.  I had to add a
"helper widget" with a clickable frame
(i.e. the white box).