Arrow keys not working on error plots
This was a big issue.  I ended up
consolidating the liveTimeNext/Prev
functionality into the book model.

Now both the curves view and live time
input in the bottom left entry box
share the same code and support error
curves.

A novel addition is the curvemodel
painterpath which takes an error
painter path and makes a CurveModel
out of it so that compare curves and
error curves use the same business logic.

Compare Curves use a cached CurveModel that is
housed in the BookModel whereas error
curves are PainterPaths generated from
two CurveModels.