Fix the alloc_type of allocations performed by declare_var. #1529
declare_var() should indicate in ALLOC_INFO::alloc_type, how
it has allocated the new object. From this we know how to delete it.

declare_var() can allocate an object in one of two ways:
1) using  malloc or calloc, or
2) using io_src_alloc_<CLASS>, which is generated by ICG.

PROBLEM 1 :
ALLOC_INFO::alloc_type was not properly set when allocating a composite type.
It should have been set to TRICK_ALLOC_NEW to indicate that it was allocated
by io_src_alloc_<CLASS>.

PROBLEM 2:
The enumeration value TRICK_ALLOC_NEW is mis-named and is misleading. The io_src_alloc_<CLASS>
routines do not necessarily allocate memory using new. If the object is a POD type, it uses calloc.

SOLUTION:
So, THIS commit:

1) Renames TRICK_ALLOC_NEW to TRICK_ALLOC_IOSRC so as not be misleading.

2) In declare_var(), sets alloc_type = TRICK_ALLOC_IOSRC when allocating a composite type (TRICK_STRUCTURED).