Introduce framework for TV param drag-n-drop
Ultimately, the main reason for drag-n-drop
is to drop a parameter onto the x-axis
of a plot to change the plot domain.

This is a step towards being able to both
drop VarsModel vars as well as params from the
SieListModel list of TV params onto a plot's
x axis to change the domain.