Drag and drop tv param onto empty tab widget
This commits makes it so When one drags and
drops an sie tv param onto an empty page in
the tab widget.

If alt is pressed when dropping, all variables
expanded from the param are added to the plot.
If ctl is pressed, individual plots are
added to the page.
If neither ctl or alt is pressed, new pages
are created from each of the variables
exanded from the sie param.