Move sie logic from TrickView to SieListModel
This is in preparation of passing the SieListModel
to the DPTreeWidget so DPs can work with Trick's
TV for live stripcharting DP spec files :)

Prior to this commit, the TrickView class was
doing all the work of building the param
list from the Trick var server SIE DB.  Now,
instead, the SieListModel encapsulates it all
and can be used by different views.  So, the SIE
Model can be used by TrickView and DPTreeWidget.