Handle tv params that are not allocated
When I tested the GIVS sim there were
params that were not allocated by
Trick, most likely new-ed or malloc-ed.

There is no way to tell whether or not
the param is allocated if not alloced
by Trick.

This commit makes it so when one clicks
a param that is not managed by Trick,
a popup will let the user know, but will
not plot the param.  If this is too
restrictive, I will need to implement a
dialogue to choose a hand-typed in
parameter with whatever dimensions the
user wants.  This is what Trick TV does.

In the case of a DP file with params
that aren't managed by Trick, it will
go ahead and try to plot them assuming the
param is allocated.  If the DP param
isn't allocated in memory, this could
crash the sim.  It's a catch 22.