Use Shift key to create new page in TV tab
Prior to this commit, one could not Ctrl click
a parameter to expand and add the parameter
to a new page.  The Ctrl click would add the
parameter to the current page.

Now, if one hold Shift+Ctrl and clicks a param,
it will add the expanded param to a new page.

Similarly, if one holds Shift+Alt when clicking
a param, the param will be expanded in a single
plot on a new page.