Protect sie parameter string list by mutex
The SIE model's parameter stringlist is
loaded in a thread.  If the _params stringlist
is not protected, a conflict occurs when
the thread is appending params to the
_params stringlist while the model trying
to access _params.

If not protected, a segmentation fault
occurs while even doing something as
simple as trying to get _params.size().