Refresh DP filter after SIE model loads
If this isn't done, the filter can occur
prior to the SIE being loaded.  If the
SIE isn't entirely loaded the filter
will prematurely filter out DPs since
the params of the DPs aren't loaded from
the SIE model.

To make it work, simply re-filter once
the SIE model is completely loaded.