TV Filter for Trick managed variables when reasonable
Prior to this change, when one entered a search in the
koviz TV search box it returned a list of params in
the SIE database regardless if those params exist
in Trick managed memory i.e. plottable params.

The best would be if the search could give only
plottable params but that isn't possible because
sometimes a search yields thousands of params
and determining if each param is in Trick managed
memory would take a formiddably long time.

To handle the case when a search result yields
too many params, I just made a cut off.  If the number of
searched for params in the koviz TV is <= 256, filter
that search further for Trick memory managed parameters.
If it is > 256, simply return the list, and perform
no extra memory check.