Reconnect to sim if sim goes down
If koviz is disconnected from a live sim,
try to reconnect to the sim.

If the sim comes back up, this will allow one
to coplot live sim connections.  This means a sim
developer can run the sim, plot, shutdown
the sim, mod the sim, bring it back up,
plot again and immediately see how the modded
sim behaves compared to the last sim run.