Do not emit dataChanged() in PlotBookModel::setData()
The dataChanged signal will be emitted in
PlotBookModel::setData() when the parent
method QStandardItemModel::setData(idx,value,role);
is called.

Emiting that signal early is redundant and also
is premature because the data at idx is actually
not yet set.  This was causing a core dump in
the TV model which was trying to replace a CurveModel
at idx.