Handle TV Model reset when sim disconnects
When the sim disconnects, the underlying
model (_params) of the TV model gets
cleared.  Before the model is reset,
curves that are using TV model copy
the TV model curve data into a new
CurveModel (CopyCurveModel) so the data
for the curve persists in the copy.

This commit introduced the CopyCurveModel.