Make curve model copies on TV disconnect
When one is viewing curves from a live
sim connection, the curve data is in the
TV model.  When the sim disconnects, this
commit replaces the curve live data with a copy.
This is done so that reconnections take
on new live data and the old connection is
safely stored in a copy of the original
live data.