Live data plots not refreshing
In this commit:
dcd9974c8bfe6ceeb9af7db87e02702771e336e1

I took out an "emit dataChanged()" signal
in the PlotBookModel thinking it was not
needed.  I then got a report that curves
were not always refreshing.  The plots
were only refreshing if the plot rectangle
changed.  When plotting time domain data,
the refresh is constant since the plot
rectangle is changing; however, with an
xy orbit where the plot rect goes static,
the curves would not update until one refreshed
the plot.

This commit has PlotBookModel::setData() do
all the work of setting the data and emitting
the dataChanged() signal.

I tested and see no core dumps.  I think this is
the way to go.

This is the code comment:

Instead of using QStandardItemModel::setData(idx,value,role)
at the bottom of this method, set the data here and emit the
dataChanged() signal.  This is necessary since calling
QStandardItemModel::setData(idx,value,role);
doesn't emit the dataChanged() signal since, I guess, the value
is an unchanging pointer.  If the dataChanged() signal is
not emitted, the views will not refresh correctly.