Add satellite symbol to pdf painter
This adds the PDF drawing of a satellite
symbol.  The PDF and X drawing code are
almost identical and really need to be
consolidated.

The problem is that the sizes aren't exactly
scalable, so I've just separated the PDF
from the X drawing.  This makes it so I can
draw in PDF exactly how I want without trying
to find some happy medium between X and PDF
drawing --- but there is too much duplication.

Anyhow, this satellite symbol was more for fun,
but does highlight the need for a common drawing
between PDF and X.