Painter path adding NaN
Koviz was giving this error msg:
QPainterPath::arcTo: Adding arc where a parameter is NaN, results are undefined

I tracked it down to a coord2pixel transform having inf and nans,
and tracked that down to the transform using an empty math
rectangle.

To fix it, don't create live pix map if the mathrect has a width of zero
or height of zero.