big: Introduce pure virtual Run class
This commit makes the way for different
Run sources other than RUN directories.
Up to this point, koviz accepted a set
of RUNs on the command line and made
DataModels from those RUNs.

A use case came up where both Kaitlin
and Sanjay wanted to run koviz on
a set of files rather than a set
of directories, like so:

    % koviz *.csv

This commit abstracts what a Run
is and makes the way for making
a koviz "Run" from a file rather
than a RUN_directory.

This commit adds no new functionality,
it only adds the Run class abstraction and
virtualizes a Run directory into a
RunDir class that implements the pure
Run abstract class.