Curve marker annotations need run path
Prior to this commit, if on clicks a
curve on a plot that has more than 5
curves, the marker would contain the
RUN directory.  Since RUNs are now
abstracted and can be files instead
of RUN directories, the annotation
needs the path which could be a RUN dir,
 a file or even a Trick sim.

This commit adds the RUN path to the
DataModel class so affected a lot of
files.