Fix core on live sim shutdown
While viewing live data with koviz
then shutting the sim, koviz was
seg faulting on a call to get the
run path for the legend.  Turns out
that the TV data model was being deleted
when the sim went down, so needed
to make sure the tv model wasn't null
before accessing it.