Do not wait on Blender connection
After the WinKoviz release, there were two reports
that koviz was "laggy" after selecting a curve.

I replicated the issue and found that koviz was
attempting to connect to and send time to a Blender
server (and "the visualizer").

It was doing this while the mouse moved, so was
repeatedly trying to connect.  The connect had
a timeout which was being honored on the Windows
side and making mouse movement laggy.

To keep the koviz <-> Blender auto connection,
I removed the timeout.  So, this commit keeps the
auto connect, which means you can bring Blender
(and/or koviz) up/down at will and the two will
connect.

I say "autoconnect", if Blender goes down while
koviz is up, one needs to deselect/reselect curve on koviz
to regain connection.