WinKoviz username is empty
I was grabbing the username from the
current directory's file info owner.
On Windows that doesn't exist, so changed
getting the username to getting the
home directory name.  This works on all
platforms.

This is a little different than the past, in the past,
I used the username of the directory which could have
been a different user.  Now, it's just the username of
whoever is running koviz.