Run path too long in live curve annotation
When one plots a bunch of RUNs and then selects
one of the curves, the annotation was putting
the run path in the live coord arrow annotation.
This path was too long in most cases.  In some cases,
if there wasn't enough room to annotate, the
annotation didn't appear at all.

To get around this, I simply show the RUN_#####
dir name if a RUN_#### is in the path, but just
"RUN_#####".  This makes the annotation short.

If one clicks a curve the entire curve path
is in the koviz status at the bottom of the
koviz window, so the RUN_name annotation is
more of an addition anyhow.