Introduce RUN refresh
This is the first in a series of commits
that introduce refreshing RUNs.

"Refreshing RUNs" will give the ability
to click the Refresh button and have all
plot data reloaded.  This will enable sim
modelers the capability to run a sim, look
at the data, tweak the sim, rerun the sim,
then refresh the plotted data with the
new sim run.

This will hopefully help engineers with
the iterative process of sim building.