Run refresh not handling case when logged sim vars change
This handles the case when one runs koviz on a Run, then
reruns the Run and logs a new set of variable names.

When refreshing Runs via the koviz menu - if there are
plots of variables up from the original Run that do not
exist in the latest Run, koviz will keep the original.