DP Table needs to check if live coord is empty
While testing tables, koviz aborted with an
error while trying to access live coord time.
Turns out, live coord time was empty, which
caused koviz to abort when trying to convert
the empty string to a double.

The fix is to simply check if live coord time
is empty before trying to access the actual
value.