Do not abbreviate user supplied DP y legend labels
Koviz normally abbreviates legend labels.  If the
user supplies a DP y label (which isn't the same
as the y variable itself!), use the user supplied
y label as is with no abbreviation.

In a lot of cases, DPs duplicate the y variable
name as the ylabel.  If the y variable matches
the y label, koviz will abbreviate.  This is to
keep long variable names from cluttering up the
plot.