Speed up search by skipping dirs which are too deep
The RunWidget regexp filter was hanging if I brought
up koviz on the root level ("/").  To get around that,
just accept the dir if it too deeply nested.  It's
bogus to accept a dir that doesn't contain the regexp,
but more bogus to recursively search the entire
directory system since it hangs.

The worst thing that happens is that deeply nested
directories always show up in the search, but
in practice, I don't think it will be an issue.