Add "ClearRuns" menu option
This option allows users to clear a current
set of RUNs and then add new RUNs.

When one clears runs, and load a new RUN,
the variable list gets updated. This enables
users to plot one RUN set, clear the plots/runs,
load a new RUN, and then click the new
variables.

I was able to peruse RUNs this way (by clearing
and adding new RUNs) on a giant set of RUNs that
have nothing to do with one another - like RUN
surfing.

This is a big step in making koviz work
non-commandline.