Add csv validation when drag-n-dropping a csv file onto RUNs
Since koviz only handles Trick formatted csv data files,
add validator so that user gets a nice popup message
if the user attempts to drag-n-drop a malformed csv onto
the RUNs list from the RUNs widget.

This commit also takes the timeNames into account.

I also added a small sanity check to ensure the number
of header columns is the same as the number of value
columns for the first data row in the csv.