Add ability to remove RUN curves
This commit adds the ability for the user
to remove a RUN (and associated curves) by
selecting a RUN in the RUNs list and then
pressing the delete key.

This is another big step towards running
koviz non-commandline.  One can add news runs,
clear runs and now delete/remove individual
runs.

Another thing this capability gives is the
following workflow:
   1. Select a run
   2. Plot vars from that run
   3. Delete the run
      This will keep empty plots up from
      the last run
   4. Select a new run (with same vars)
      Fills the empty plots with
      vars from the newly selected run

That workflow will be nice for SDTS testing.
As new runs are generated, all one will have
to do is remove the current run, and select
the new one.  Alternatively, for comparing
simply drag-n-drop the new run to the runs list.