Reduce runs filter search to a single sub-directory
After testing remotely in /users/core in the Tricklab,
I found it better to search shallow for speed sake.

The filter now only goes a single directory in depth.

In practice, I think this will be okay.  The main thing
is that the filter will keep the selection to files/dirs
that are plottable by koviz.  Second, the filtering
is much faster when the depth is 1.