Fix MonteInputsView::keyPress/dropEvent() methods
I found I had inadvertantly caused the key event
to no longer propogate because one has to call
the parent class and/or do a event->ignore() so the
key events will go to the next widget.

The symptom of the issue was plotting a monte carlo
and then clicking a row in the monte inputs, then
found clicking the down arrow did nothing, so had
inadvertantly short circuited the key press down arrow.