Remove TimeIt classes
This is the last commit in a series of commits that
remove the TimeIt classes from koviz.  I made the TimeIt
classes originally thinking there was no crossplatform
elapsed timer (or maybe it didn't exist at the time).
QElapsedTimer is exactly what I needed, so the replacement
was very straightforward and now timers work on all
platforms. Timers are used in progress dialogs to give
the user the elapsed time for long operations.