Koviz is not allowing DP file with -trickport on commandline
This is another issue due to not testing tv when I did
the dynamic drag-n-drop run addition.

The issue this fixes is the following, when running
something like:

% koviz -trickport 4545 DP_xy

koviz was aborting, saying there was no RUN, but should
need no run since plotting from the variable server data
without a run dir or run file.