Introduce WinKoviz instructions and icon
This commit includes a Windows readme
that gives instructions on how to build,
deploy and create an installer on Windows.

It also introduces the icon.  The icon was
created using koviz and the included
spring monte carlo.  The icon is a 1000 monte
carlo plot of vel vs accerlation of the damped
spring - varying the spring k value.

The negative space of the icon is a wave.
WinKoviz started in Waikiki after a nut from
a palm tree fell and struck me.  Not as earth
shattering as the law of gravity, but hopefully
users will find some Aha moments perusing
plots on Windows.  Good luck!