Clean up plot tic close to zero
On 8/8/2024 on this commit:
8adaf0e6a8b77e338b81c27e2479799699a0ee47
I took out logic that formatted tics
close to zero as zero and instead used
the actual tic value no matter how
close to zero.

This introduced another issue
when tic values were close to zero
but actually meant to be zero.

For example if Tics={-1e-17,1e-34,1e-17},
the plot should really have
Tics={-1e-17,0,1e-17}.

This commit fixes that issue by replacing
the tic closest to zero with zero if
appropriate.