Realtime report needs update for Trick job thread id
I was taking a job name like:
    JOB_thread_process_event_3.ep.process_event_C3.266.03
and parsing out the thread ID from the name.  I assumed Trick's
naming convention added the CX for the thread ID.  I think Trick
used to adhere to that, but maybe I made a bad assumption.
This issue came up because there were jobs with the _CX.num.num
suffix where the X was NOT the thread ID.

Instead of getting the threadID from the job name,
use the filename the job comes from. For example,
if the job comes from:
    log_trick_frame_userjobs_C5.trk
it belongs to thread 5.

Another example is the following job name:
    JOB_thread_process_event_3.ep.process_event_C3.266.03(automatic_1.000)
which came from file:
    log_frame_trickjobs.trk
Even though the job has "C3.266.03", it's not in child thread 3,
it's in thread 0, since log_frame_trickjobs.trk logs jobs in
thread 0.