Add HLine label placement/orientation
Prior to this commit, the HLine label was
drawn on the left side of the plot under/over
the hline.  This commit allows for placement
of the hline label.  The reason for placement
is that the label could cover up plot curves.
The placement allows one to put the label "left",
"center" or "right" - and allows a placement
with a value between 0 and 1, where 0 is left,
and 1 is right.  A value like 0.75 would put the
label 3/4 way down the hline.